Move product route to another file
