--- babel.config.cjs
@@ -1,1 +1,0 @@
-module.exports = {presets: ['@babel/preset-env']}
