+++ src/index.js
@@ -1,184 +1,70 @@
 import express from "express";
 import cors from "cors";
 import mongoose from "mongoose";
-import userRouter from "./routes/user.routes";
+import userRouter from "./routes/user.routes.js";
 import cookieParser from "cookie-parser";
-import imageUpload from "./middlewares/imageUpload";
-import URLQueryHandler from "./middlewares/URLQueryHandler";
-import productFormHandler from "./middlewares/productFormHandler";
-import { createTable } from "./controller/clienteController";
-import clientRoute from "./routes/clientRoute";
-import {
-  getFeaturedProdutos,
-  countProdutos,
-  createProductTable,
-  deleteProduto,
-  getProdutos,
-  getProduto,
-  insertProduto,
-  updateProduto,
-} from "./controller/produtoController";
-import deleteImages from "./utils/deleteImage";
-import formatProduct from "./utils/formatProduct";
+import { createTable } from "./controller/clienteController.js";
+import clientRoute from "./routes/clientRoute.js";
+import { createProductTable } from "./controller/produtoController.js";
+import productRouter from "./routes/productRoute.js";
 import {
   createPedidoTable,
   deletePedido,
   getAllPedidos,
   getPedidosFromClient,
   insertPedido,
-} from "./controller/pedidoController";
+} from "./controller/pedidoController.js";
 import {
   createProdutoPedidoTable,
   getAllProdutosFromPedido,
   updatePedido,
-} from "./controller/produtoPedidoController";
-
-mongoose.connect(
-  "mongodb+srv://admin:admin@cluster0.1bfhxjk.mongodb.net/?retryWrites=true&w=majority"
-);
+} from "./controller/produtoPedidoController.js";
 
 const app = express();
 
-app.use(express.json());
-app.use(cors({
+
+function configureApp() {
+  app.use(express.json());
+  app.use(cors({
     credentials: true,
     origin: 'http://localhost:3000',
     allowedHeaders: ['Content-Type']
-}))
-app.use(cookieParser());
-const routers = [
-  userRouter,
-  clientRoute
-]
+  }))
+  app.use(cookieParser());
+  const routers = [
+    userRouter,
+    clientRoute,
+    productRouter
+  ]
 
-for (const router of routers) {
-  app.use(router);
+  for (const router of routers) {
+    app.use(router);
+  }
+
+  // Usado pelas tags <img> no HTML para mostrar as imagens salvas
+  app.use("/images", express.static("src/public/images"));
 }
 
-// Usado pelas tags <img> no HTML para mostrar as imagens salvas
-app.use("/images", express.static("src/public/images"));
+async function configureMongoose() {
+  try {
+    await mongoose.connect(
+      "mongodb+srv://admin:admin@cluster0.1bfhxjk.mongodb.net/?retryWrites=true&w=majority"
+    );
+  } catch (err) {
+    console.log(err);
+  }
+}
 
-Promise.all([createTable(), createProductTable(), createPedidoTable(), createProdutoPedidoTable()]);
+async function createAllTables() {
+  Promise.all([createTable(), createProductTable(), createPedidoTable(), createProdutoPedidoTable()]);
+}
+
+configureApp();
+await configureMongoose();
+await createAllTables();
 
 app.get("/", (req, res) => {
   res.send("Bem vindo ao nosso Projeto :)");
-});
-
-/**
- * TODO:
- *      -> Fazer verificações relacionadas à segurança;
- *      -> Criar autenticação para as operações de post, put e delete.
- */
-
-// A quantidade de resultados é enviada para o front-end para facilitar a criação dos botões de paginação
-app.get("/produtos", URLQueryHandler, async (req, res) => {
-  try {
-    const result = await getProdutos(req.query);
-    const numberOfResults = await countProdutos(req.query);
-    const produtos = result.map(formatProduct);
-    if (produtos.length > 0) {
-      res.send({
-        products: produtos,
-        results: numberOfResults.size,
-        pages: Math.ceil(numberOfResults.size / req.query.pageSize),
-        currentPage: req.query.page,
-      });
-    } else {
-      res.status(204).send();
-    }
-  } catch (error) {
-    console.log(error);
-    res.status(204).send();
-  }
-});
-
-app.get("/produto/:slug", async (req, res) => {
-  try {
-    const result = await getProduto(req.params.slug);
-    if (result) {
-      const produto = formatProduct(result);
-      res.status(200).send(produto);
-    } else {
-      res.status(204).send();
-    }
-  } catch (error) {
-    console.log(error);
-    res.status(204).send();
-  }
-});
-
-app.get("/destaques", async (req, res) => {
-  try {
-    const result = await getFeaturedProdutos();
-    const produtos = result.map(formatProduct);
-    if (produtos.length > 0) {
-      res.status(200).send({ products: produtos });
-    } else {
-      res.status(204).send();
-    }
-  } catch (error) {
-    console.log(error);
-    res.status(204).send();
-  }
-});
-
-app.post(
-  "/novo-produto",
-  imageUpload.array("images", 5),
-  productFormHandler,
-  async (req, res) => {
-    try {
-      const images = req.files.map((img) => img.filename);
-      await insertProduto({ ...req.body, images: images });
-      res.status(201).send({errors: null});
-    } catch (error) {
-      console.log(error);
-      res.status(418).send({errors: {
-        nameError: 'erro',
-        priceError: 'erro',
-        descriptionError: 'erro',
-      }});
-    }
-  }
-);
-
-// TODO: testar o method put
-app.put(
-  "/produto/:slug",
-  imageUpload.array("images", 5),
-  productFormHandler,
-  async (req, res) => {
-    try {
-      const produto = await getProduto(req.params.slug);
-      if (produto) {
-        const images = req.files.map((img) => img.filename);
-        deleteImages(JSON.parse(produto));
-        await updateProduto({ ...req.body, images: images });
-        res.status(200).send("Produto atualizado");
-      } else {
-        res.status(204).send();
-      }
-    } catch (error) {
-      console.log(error);
-      res.status(204).send();
-    }
-  }
-);
-
-app.delete("/produto/:slug", async (req, res) => {
-  try {
-    const produto = await getProduto(req.params.slug);
-    if (produto) {
-      await deleteProduto(req.params.slug);
-      deleteImages(JSON.parse(produto.images));
-      res.status(200).send("Produto deletado");
-    } else {
-      res.status(204).send();
-    }
-  } catch (error) {
-    console.log(error);
-    res.status(204).send();
-  }
 });
 
 app.get("/pedido", async (req, res) => {
